+++ lib/BullsEyeServer/src/Routing/RoutingExceptions.dart
@@ -2,26 +2,22 @@ part of softhai.bulls_eye.Server;
 
 abstract class RoutingException extends HttpRequestException {
 
-  Route route;
+  common.Url url;
   
-  RoutingException(HttpRequest request, this.route) : super(request);
+  RoutingException(HttpRequest request, this.url) : super(request);
   
   String toString() {
-    var routeName = this.route._routeDefenition.name;
-    
-    return super.toString() + "The route '$routeName' creates an unhandled error";
+    return super.toString() + "The route '${this.url.name}' creates an unhandled error";
   }
 }
 
 class FileNotFoundException extends NotFoundException {
 
-  Route route;
+  common.Url url;
   
-  FileNotFoundException(HttpRequest request, String resource, this.route) : super(request, resource, "File");
+  FileNotFoundException(HttpRequest request, String resource, this.url) : super(request, resource, "File");
   
   String toString() {
-    var routeName = this.route._routeDefenition.name;
-    
-    return super.toString() + " This was reported by the route '$routeName'.";
+    return super.toString() + " This was reported by the route '${this.url.name}'.";
   }
 }